common: Forcibly include customization.mk at the beginning of common.mk
This resolves two subtle issues with the inherit-product mechanism:
- Inherited files are evaluated later, at least after _common.mk_ has
  been fully read despite the call to inherit-product being at the top
  of the file;
- The ordering between other inherit-product calls is not guaranteed;
  the list is sorted to remove duplicates and provide consistency.

The observed problem at hand is overriding TARGET_USE_QCRILD in
customization.mk, as is happening in PE. Due to sorting common-init.mk
(which uses this variable) is evaluated _before_ customization.mk,
leading to the wrong rc files ending up on vendor.img.
CommonConfig.mk on the other hand is parsed after the entire product
inheritance tree is evaluated, hence the right vintf is selected
according to this variable.

This issue is fixed by forcefully including customization.mk at the
"right" (or at the very least expected) step in the build process.
Repository maintainers are aware that this file should only be included
once, which is safe enough to do away with the single-evaluation
guarantees inherit-product provides.

But why is that "right" place at the bottom of common.mk?
As explained above, inherited files are evaluated _at least_ after the
current file (common.mk) has been fully read meaning that all variables
are available in customization.mk.
Writers of customization.mk have come to rely on these variables being
available, such as PE expecting COMMON_PATH to be set.

Source:
https://android.googlesource.com/platform/build/+/android-10.0.0_r26/core/product.mk#364
> # To be called from product makefiles, and is later evaluated during
> # the import-nodes call below. It does three things:

Signed-off-by: Marijn Suijten <marijns95@gmail.com>